requirements.txt
django
djangorestframework
django-cors-headers
drf-spectacular
django-rest-swagger
paramiko